Move this into the main file, it's more practical
This wasn't entirely necessary, but it also didn't make any sense why
this was in an oddly named separate file in this case, so I moved it.
